requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.1
pyasn1==0.4.8
Pygments==2.5.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
